Store Gree coodinators in entry runtime_data
