+++ homeassistant/components/gree/const.py
@@ -2,15 +2,12 @@
 
 COORDINATORS = "coordinators"
 
-DATA_DISCOVERY_SERVICE = "gree_discovery"
-
 DISCOVERY_SCAN_INTERVAL = 300
 DISCOVERY_TIMEOUT = 8
 DISPATCH_DEVICE_DISCOVERED = "gree_device_discovered"
 DISPATCHERS = "dispatchers"
 
 DOMAIN = "gree"
-COORDINATOR = "coordinator"
 
 FAN_MEDIUM_LOW = "medium low"
 FAN_MEDIUM_HIGH = "medium high"
