+++ homeassistant/components/gree/switch.py
@@ -13,12 +13,12 @@ from homeassistant.components.switch import (
     SwitchEntity,
     SwitchEntityDescription,
 )
-from homeassistant.config_entries import ConfigEntry
 from homeassistant.core import HomeAssistant, callback
 from homeassistant.helpers.dispatcher import async_dispatcher_connect
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
-from .const import COORDINATORS, DISPATCH_DEVICE_DISCOVERED, DOMAIN
+from . import GreeConfigEntry
+from .const import DISPATCH_DEVICE_DISCOVERED
 from .entity import GreeEntity
 
 
@@ -92,7 +92,7 @@ GREE_SWITCHES: tuple[GreeSwitchEntityDescription, ...] = (
 
 async def async_setup_entry(
     hass: HomeAssistant,
-    entry: ConfigEntry,
+    entry: GreeConfigEntry,
     async_add_entities: AddEntitiesCallback,
 ) -> None:
     """Set up the Gree HVAC device from a config entry."""
@@ -106,7 +106,7 @@ async def async_setup_entry(
             for description in GREE_SWITCHES
         )
 
-    for coordinator in hass.data[DOMAIN][COORDINATORS]:
+    for coordinator in entry.runtime_data:
         init_device(coordinator)
 
     entry.async_on_unload(
