+++ tests/components/gree/conftest.py
@@ -20,7 +20,7 @@ def mock_setup_entry() -> Generator[AsyncMock]:
 @pytest.fixture(autouse=True, name="discovery")
 def discovery_fixture():
     """Patch the discovery object."""
-    with patch("homeassistant.components.gree.coordinator.Discovery") as mock:
+    with patch("homeassistant.components.gree.Discovery") as mock:
         mock.return_value = FakeDiscovery()
         yield mock
 
@@ -29,7 +29,7 @@ def discovery_fixture():
 def device_fixture():
     """Patch the device search and bind."""
     with patch(
-        "homeassistant.components.gree.coordinator.Device",
+        "homeassistant.components.gree.Device",
         return_value=build_device_mock(),
     ) as mock:
         yield mock
